reuse content preview for requests
